using powerselect

+++ app/components/job-applications/compact.js
@@ -5,7 +5,13 @@ export default class JobApplicationsCompact extends Component {
   @service flashMessages;
   @service router;
 
+  get jobApplication() {
+    return this.args.jobApplication;
+  }
   @action delete() {
+    const message = `${this.jobApplication.get('jobPost.title')}`;
+    const allowed = confirm(`Are you sure you want to delete ${message}`);
+    if (!allowed) { return }
     this.args.jobApplication
       .destroyRecord()
       .then((record) => {
